backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
redis==5.0.4
pymongo==4.7.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
motor==3.4.0
sqlalchemy==2.0.30
aiofiles==23.2.1
python-multipart==0.0.9
